Don't let a class sweep up a field that doesn't yet exist.
If you try to do this, your method won't actually be swept up, and
you'll be confused. So fail early and loudly.

(And actually, since we're redefining a method that doesn't yet exist,
Ruby helps us out with a NameError anyway, so this is really about
providing a clearer error message.)